cli/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[versions]
stable_version = "0.0.3"
current_version = "0.0.4"

[tool.poetry]
name = "cli"
version = "0.0.4"
description = "DevOps Utility CLI"
authors = [
  "Tyler Peckenpaugh <peckenpaugh@gmail.com>",
  "Justine Ginchereau <justineginchereau@gmail.com>"
]
readme = "../README.md"

[tool.poetry.scripts]
dev = "cli.main:app"

[tool.poetry.dependencies]
python = ">=3.9,<4"
pydantic = "^1.9.1"
boto3 = "^1.24.34"
configparser = "^5.2.0"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pytest_mock = "^3.7.0"
pytest-freezegun = "^0.4.0"
pytest-dotenv = "^0.5.2"
toml = "^0.10.2"

[tool.poetry.dependencies.typer]
extras = ["all"]
version = "^0.6.1"

[tool.poetry.dev-dependencies.moto]
extras = ["ssm", "s3"]
version = "^3.1.18"
